extract large test
