+++ ydb/core/tx/schemeshard/ut_login_large/ya.make
@@ -1,0 +1,20 @@
+UNITTEST_FOR(ydb/core/tx/schemeshard)
+
+FORK_SUBTESTS()
+
+SIZE(LARGE)
+TAG(ya:fat)
+
+PEERDIR(
+    ydb/core/testlib/default
+    ydb/core/tx
+    ydb/core/tx/schemeshard/ut_helpers
+)
+
+YQL_LAST_ABI_VERSION()
+
+SRCS(
+    ut_login_large.cpp
+)
+
+END()
